Handle non-200-level responses without throwing
See #6 for more context. If you are making an HTTP request, don't throw
unless there is a syntax error or a network connection error. If the
server responds with a 404 or a 501 or anything else, the request
succeeds and it is the responsibility of the client to interpret the
value of `response.status` and react accordingly.