stop using babel-polyfill but keep supporting node 4 (and old browsers)
json-fetch was using some methods that are part of the ES2015 standard libary
(Array.includes and String.includes). For node 4 support, it had pulled in the
babel-polyfill node module. Unfortunately this breaks support for projects that
already include babel-polyfill, since it can only be included once.

This commit fixes this by removing the ES2015 calls and replacing them with
ponyfills (the array-includes and string-includes npm modules).